Add 'clippy' and 'rustfmt' to the build process.
This ensures that on save, source files are automatically reformatted
according to rustfmt rules.

Clippy warnings are shown before building. (But the build happens
regardless). Besides this devs might add their own in-editor clippy
config.

Clippy currently ignores the `renderer` module and its submodules,
because the `grrrlib` source is included inline in that file.

It might be cleaner to separate this further, so we truly only ignore
the generated unsafe code, while still checking the main rendering logic.
(Clippy complains about missing safety comments for each of the unsafe
functions :shrug: )